Removed unused files, mvp refactoring
