+++ app/src/main/java/com/chichkanov/yandex_weather/ui/weather/WeatherFragment.java
@@ -1,8 +1,10 @@
-package com.chichkanov.yandex_weather;
+package com.chichkanov.yandex_weather.ui.weather;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+
+import com.chichkanov.yandex_weather.R;
 
 public class WeatherFragment extends Fragment {
